Giving up on encryption
and life

+++ src/routes/api/admin/guilds/[guild]/categories/[category].js
@@ -1,10 +1,14 @@
 const { logAdminEvent } = require('../../../../../../lib/logging');
+const { randomUUID } = require('crypto');
 
 module.exports.delete = fastify => ({
 	handler: async (req, res) => {
 		/** @type {import('client')} */
 		const client = res.context.config.client;
+		const guildId = req.params.guild;
 		const categoryId = Number(req.params.category);
+		const original = req.params.category && await client.prisma.category.findUnique({ where: { id: categoryId } });
+		if (!original || original.guildId !== guildId) return res.status(404).send(new Error('Not Found'));
 		const category = await client.prisma.category.delete({ where: { id: categoryId } });
 
 		logAdminEvent(client, {
@@ -27,12 +31,13 @@ module.exports.get = fastify => ({
 	handler: async (req, res) => {
 		/** @type {import('client')} */
 		const client = res.context.config.client;
+		const guildId = req.params.guild;
 		const categoryId = Number(req.params.category);
 		const category = await client.prisma.category.findUnique({
 			include: {
 				questions: {
 					select: {
-						createdAt: true,
+						// createdAt: true,
 						id: true,
 						label: true,
 						maxLength: true,
@@ -45,8 +50,10 @@ module.exports.get = fastify => ({
 					},
 				},
 			},
-			where: { id: categoryId  },
+			where: { id: categoryId },
 		});
+
+		if (!category || category.guildId !== guildId) return res.status(404).send(new Error('Not Found'));
 
 		return category;
 	},
@@ -57,11 +64,13 @@ module.exports.patch = fastify => ({
 	handler: async (req, res) => {
 		/** @type {import('client')} */
 		const client = res.context.config.client;
+		const guildId = req.params.guild;
 		const categoryId = Number(req.params.category);
 		const guild = client.guilds.cache.get(req.params.guild);
 		const data = req.body;
 		const original = req.params.category && await client.prisma.category.findUnique({ where: { id: categoryId } });
-		if (!original) return res.status(404);
+		if (!original || original.guildId !== guildId) return res.status(404).send(new Error('Not Found'));
+
 
 		if (data.hasOwnProperty('id')) delete data.id;
 		if (data.hasOwnProperty('createdAt')) delete data.createdAt;
@@ -70,11 +79,14 @@ module.exports.patch = fastify => ({
 			data: {
 				...data,
 				questions: {
-					upsert: data.questions?.map(q => ({
-						create: q,
-						update: q,
-						where: { id: q.id },
-					})),
+					upsert: data.questions?.map(q => {
+						if (!q.id) q.id = randomUUID();
+						return {
+							create: q,
+							update: q,
+							where: { id: q.id },
+						};
+					}),
 				},
 			},
 			where: { id: categoryId },
